asyncapi/rest: extract JsonSchemaToGeneConverter behind a RefResolver (M4c)
Removes the synthetic-OpenAPI bridge AsyncAPI was using to call into
RestActionBuilderV3.getGene.  The bridge made an AsyncAPI build look
like an OpenAPI parse just to satisfy getGene's REST-shaped parameter
list — that worked but was an obvious smell, and the abstraction
debt was going to compound when M5 brought white-box AsyncAPI online.

What changed:

- New search/gene/builder package:
  - SchemaRefResolver: protocol-agnostic interface, takes a JSON-Schema
    `$ref` and returns a Swagger Schema (or null).
  - JsonSchemaToGeneConverter: holds the ~800 LOC of JSON-Schema → Gene
    logic that used to live as private helpers on RestActionBuilderV3
    (getGene, getNonNullGene, createObjectGene, allOf/anyOf/oneOf
    assembly, constraint handling, refCache, example/default machinery,
    string/number/array/object builders).  Takes a SchemaRefResolver +
    Options at construction; no REST types in its surface.
- REST: RestSchemaRefResolver wraps RestSchema/SchemaOpenAPI and
  delegates to SchemaUtils.getReferenceSchema.  RestActionBuilderV3
  now constructs one converter per build (or per DTO call), threads
  it through handlePathItem → handleOperation → extractParams →
  handleParam / handleBodyPayload, and calls converter.getGene(...)
  instead of the old in-place pipeline.  The 1340 LOC of moved
  helpers are gone from RestActionBuilderV3; refCache moved with
  them; dtoCache stays (REST-specific synth-OpenAPI cache).
- AsyncAPI: AsyncAPISchemaRefResolver wraps the AsyncAPISchema
  componentSchemas map (JsonNode-backed) and lazily converts via
  JsonSchemaConverter on lookup.  AsyncAPIActionBuilder no longer
  synthesises an OpenAPI YAML doc, no longer calls into
  RestActionBuilderV3, and no longer rewrites $refs through identical
  prefixes — it builds a converter, hands it the AsyncAPI resolver,
  and calls converter.getGene directly.  JsonSchemaConverter survives
  but only as a thin JsonNode → Swagger Schema bridge; the doc
  comment is updated to reflect the new role.

Behavioural net: zero change.  All 342 tests across
RestActionBuilderV3Test (152), EMConfigTest (108), AsyncAPIAccessTest
(4), and the surrounding REST resource/structure/DTO tests still
pass with the same skip counts.

The follow-up the prior PR description flagged is now done.
