+++ core/src/main/kotlin/org/evomaster/core/problem/asyncapi/builder/AsyncAPIActionBuilder.kt
@@ -1,23 +1,18 @@
 package org.evomaster.core.problem.asyncapi.builder
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import io.swagger.v3.oas.models.media.Schema
 import org.evomaster.core.EMConfig
 import org.evomaster.core.problem.asyncapi.data.AsyncAPIAction
 import org.evomaster.core.problem.asyncapi.param.AsyncAPIParam
+import org.evomaster.core.problem.asyncapi.schema.AsyncAPIChannel
 import org.evomaster.core.problem.asyncapi.schema.AsyncAPIMessage
 import org.evomaster.core.problem.asyncapi.schema.AsyncAPIOperation
 import org.evomaster.core.problem.asyncapi.schema.AsyncAPISchema
 import org.evomaster.core.problem.asyncapi.schema.JsonSchemaConverter
-import org.evomaster.core.problem.rest.builder.RestActionBuilderV3
-import org.evomaster.core.problem.rest.schema.OpenApiAccess
-import org.evomaster.core.problem.rest.schema.RestSchema
-import org.evomaster.core.problem.rest.schema.SchemaLocation
-import org.evomaster.core.problem.rest.schema.SchemaLocationType
-import org.evomaster.core.problem.rest.schema.SchemaOpenAPI
+import org.evomaster.core.problem.asyncapi.schema.ReplyBinding
 import org.evomaster.core.remote.SutProblemException
 import org.evomaster.core.search.gene.ObjectGene
+import org.evomaster.core.search.gene.builder.JsonSchemaToGeneConverter
 import org.evomaster.core.search.gene.wrapper.NullableGene
 import org.slf4j.LoggerFactory
 import java.util.UUID
@@ -25,13 +20,11 @@ import java.util.UUID
 /**
  * Translates a parsed [AsyncAPISchema] into [AsyncAPIAction]s ready for sampling.
  *
- * The trick that lets us avoid duplicating REST's ~800 LOC schema-to-gene
- * pipeline: we synthesise a tiny OpenAPI document whose `components.schemas`
- * reproduces the AsyncAPI component schemas, hand it to the swagger parser,
- * and call [RestActionBuilderV3.getGene] (recently bumped to `internal` for
- * exactly this reuse).  Inline payloads are converted to swagger
- * [io.swagger.v3.oas.models.media.Schema] via [JsonSchemaConverter] and fed in
- * the same way; `$ref`s resolve through the synthetic OpenAPI components.
+ * The shared [JsonSchemaToGeneConverter] does all the JSON-Schema → Gene
+ * heavy lifting; we just feed it converted Swagger [Schema] objects (built
+ * from AsyncAPI message payloads via [JsonSchemaConverter]) plus an
+ * [AsyncAPISchemaRefResolver] that knows how to look up component schemas
+ * by `$ref`.  No synthetic OpenAPI document is constructed.
  *
  * Black-box only for now: the builder ignores reply correlation header
  * generation (the fitness layer fills in a per-evaluation UUID at publish time).
@@ -41,8 +34,6 @@ class AsyncAPIActionBuilder(private val config: EMConfig) {
     companion object {
         private val log = LoggerFactory.getLogger(AsyncAPIActionBuilder::class.java)
     }
-
-    private val yamlMapper = ObjectMapper(YAMLFactory())
 
     data class Built(
         /**
@@ -56,27 +47,20 @@ class AsyncAPIActionBuilder(private val config: EMConfig) {
 
     fun build(schema: AsyncAPISchema): Built {
 
-        val syntheticOpenApi = synthesiseOpenAPI(schema)
-        val syntheticSchema = OpenApiAccess.parseOpenApi(
-            syntheticOpenApi,
-            SchemaLocation("asyncapi://synthetic", SchemaLocationType.RESOURCE)
+        val resolver = AsyncAPISchemaRefResolver(schema.componentSchemas)
+        val converter = JsonSchemaToGeneConverter(
+            resolver,
+            JsonSchemaToGeneConverter.Options(config)
         )
-        val restSchema = RestSchema(syntheticSchema)
-        val options = RestActionBuilderV3.Options(config)
-        val messages = mutableListOf<String>()
 
         val out = LinkedHashMap<String, List<AsyncAPIAction>>()
 
         schema.operations.values.forEach { op ->
             try {
-                out[op.name] = buildForOperation(op, schema, restSchema, syntheticSchema, options, messages)
+                out[op.name] = buildForOperation(op, schema, converter)
             } catch (e: Exception) {
                 log.warn("Skipping AsyncAPI operation '{}': {}", op.name, e.message)
             }
-        }
-
-        if (messages.isNotEmpty() && log.isDebugEnabled) {
-            messages.forEach { log.debug("AsyncAPI gene-build message: {}", it) }
         }
 
         return Built(out)
@@ -85,10 +69,7 @@ class AsyncAPIActionBuilder(private val config: EMConfig) {
     private fun buildForOperation(
         op: AsyncAPIOperation,
         schema: AsyncAPISchema,
-        restSchema: RestSchema,
-        syntheticSchema: SchemaOpenAPI,
-        options: RestActionBuilderV3.Options,
-        messages: MutableList<String>
+        converter: JsonSchemaToGeneConverter
     ): List<AsyncAPIAction> {
 
         if (op.action != AsyncAPIOperation.Action.SEND) {
@@ -118,10 +99,7 @@ class AsyncAPIActionBuilder(private val config: EMConfig) {
             message = publishMessage,
             pairId = pairId,
             schema = schema,
-            restSchema = restSchema,
-            syntheticSchema = syntheticSchema,
-            options = options,
-            messages = messages,
+            converter = converter,
             replyBinding = op.reply,
             correlationLocation = publishMessage.correlationLocation
         )
@@ -149,10 +127,7 @@ class AsyncAPIActionBuilder(private val config: EMConfig) {
                     message = replyMessage,
                     pairId = pairId,
                     schema = schema,
-                    restSchema = restSchema,
-                    syntheticSchema = syntheticSchema,
-                    options = options,
-                    messages = messages,
+                    converter = converter,
                     replyBinding = null,
                     correlationLocation = replyMessage.correlationLocation
                 )
@@ -170,24 +145,23 @@ class AsyncAPIActionBuilder(private val config: EMConfig) {
         message: AsyncAPIMessage,
         pairId: String,
         schema: AsyncAPISchema,
-        restSchema: RestSchema,
-        syntheticSchema: SchemaOpenAPI,
-        options: RestActionBuilderV3.Options,
-        messages: MutableList<String>,
-        replyBinding: org.evomaster.core.problem.asyncapi.schema.ReplyBinding?,
+        converter: JsonSchemaToGeneConverter,
+        replyBinding: ReplyBinding?,
         correlationLocation: String?
     ): AsyncAPIAction {
 
         val swaggerSchema = resolvePayloadSchema(message, schema)
-        val gene = RestActionBuilderV3.getGene(
+        val messages = mutableListOf<String>()
+        val gene = converter.getGene(
             name = AsyncAPIAction.PAYLOAD_PARAM,
             schema = swaggerSchema,
-            schemaHolder = restSchema,
-            currentSchema = syntheticSchema,
             referenceClassDef = null,
-            options = options,
             messages = messages
         )
+        if (messages.isNotEmpty() && log.isDebugEnabled) {
+            messages.forEach { log.debug("AsyncAPI gene-build message: {}", it) }
+        }
+
         val unwrapped = if (gene is NullableGene) gene.gene else gene
         val payloadGene = unwrapped as? ObjectGene
             ?: ObjectGene(AsyncAPIAction.PAYLOAD_PARAM, listOf(gene))
@@ -218,18 +192,12 @@ class AsyncAPIActionBuilder(private val config: EMConfig) {
         return JsonSchemaConverter.convert(node)
     }
 
-    private fun pickFirstMessageId(op: AsyncAPIOperation, channel: org.evomaster.core.problem.asyncapi.schema.AsyncAPIChannel): String? {
-        // Operation-level messages reference channel-level message keys, but
-        // channels carry component-level message ids directly.  Take the
-        // intersection ordered by op-level appearance, falling back to channel.
+    private fun pickFirstMessageId(op: AsyncAPIOperation, channel: AsyncAPIChannel): String? {
         val opIds = op.messageIds.toSet()
         return channel.messageIds.firstOrNull { it in opIds || opIds.isEmpty() }
     }
 
-    private fun pickFirstReplyMessageId(
-        reply: org.evomaster.core.problem.asyncapi.schema.ReplyBinding,
-        channel: org.evomaster.core.problem.asyncapi.schema.AsyncAPIChannel
-    ): String? {
+    private fun pickFirstReplyMessageId(reply: ReplyBinding, channel: AsyncAPIChannel): String? {
         val opIds = reply.messageIds.toSet()
         return channel.messageIds.firstOrNull { it in opIds || opIds.isEmpty() }
     }
@@ -240,27 +208,5 @@ class AsyncAPIActionBuilder(private val config: EMConfig) {
         val idx = location.indexOf(marker)
         if (idx < 0) return null
         return location.substring(idx + marker.length).trim().takeIf { it.isNotEmpty() }
-    }
-
-    /**
-     * Builds a tiny OpenAPI 3.0 YAML document whose `components.schemas`
-     * mirrors the AsyncAPI document's, so swagger-parser can resolve `$ref`s
-     * for free.  No paths are declared; `RestSchema.validate()` is **not**
-     * called on the result because the synthetic doc has no operations.
-     */
-    private fun synthesiseOpenAPI(schema: AsyncAPISchema): String {
-        val componentSchemas = schema.componentSchemas
-        val openapiRoot = mutableMapOf<String, Any>(
-            "openapi" to "3.0.3",
-            "info" to mapOf("title" to "asyncapi-bridge", "version" to "1.0.0"),
-            "paths" to emptyMap<String, Any>()
-        )
-        if (componentSchemas.isNotEmpty()) {
-            openapiRoot["components"] = mapOf(
-                "schemas" to componentSchemas
-                    .mapValues { (_, v) -> yamlMapper.treeToValue(v, Any::class.java) }
-            )
-        }
-        return yamlMapper.writeValueAsString(openapiRoot)
     }
 }
