+++ core/src/main/kotlin/org/evomaster/core/problem/rest/builder/RestActionBuilderV3.kt
@@ -1,27 +1,20 @@
 package org.evomaster.core.problem.rest.builder
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.node.ObjectNode
 import io.swagger.parser.OpenAPIParser
 import io.swagger.v3.oas.models.OpenAPI
+import org.evomaster.core.EMConfig
 import io.swagger.v3.oas.models.Operation
 import io.swagger.v3.oas.models.PathItem
 import io.swagger.v3.oas.models.examples.Example
-import io.swagger.v3.oas.models.media.ArraySchema
-import io.swagger.v3.oas.models.media.JsonSchema
 import io.swagger.v3.oas.models.media.MediaType
-import io.swagger.v3.oas.models.media.ObjectSchema
 import io.swagger.v3.oas.models.media.Schema
 import io.swagger.v3.oas.models.parameters.Parameter
 import io.swagger.v3.oas.models.responses.ApiResponse
 import org.evomaster.client.java.instrumentation.shared.ClassToSchemaUtils.OPENAPI_COMPONENT_NAME
 import org.evomaster.client.java.instrumentation.shared.ClassToSchemaUtils.OPENAPI_SCHEMA_NAME
-import org.evomaster.client.java.instrumentation.shared.TaintInputName
-import org.evomaster.core.EMConfig
 import org.evomaster.core.Lazy
-import org.evomaster.core.StaticCounter
 import org.evomaster.core.logging.LoggingUtil
-import org.evomaster.core.parser.RegexHandler
 import org.evomaster.core.problem.api.param.Param
 import org.evomaster.core.problem.rest.data.Endpoint
 import org.evomaster.core.problem.rest.data.HttpVerb
@@ -35,29 +28,18 @@ import org.evomaster.core.problem.rest.schema.SchemaOpenAPI
 import org.evomaster.core.problem.rest.schema.SchemaUtils
 import org.evomaster.core.problem.util.ActionBuilderUtil
 import org.evomaster.core.search.action.Action
-import org.evomaster.core.search.gene.*
-import org.evomaster.core.search.gene.collection.*
-import org.evomaster.core.search.gene.datetime.DateGene
-import org.evomaster.core.search.gene.datetime.DateTimeGene
-import org.evomaster.core.search.gene.datetime.FormatForDatesAndTimes
-import org.evomaster.core.search.gene.datetime.TimeGene
-import org.evomaster.core.search.gene.numeric.*
-import org.evomaster.core.search.gene.wrapper.ChoiceGene
+import org.evomaster.core.search.gene.Gene
+import org.evomaster.core.search.gene.ObjectGene
+import org.evomaster.core.search.gene.builder.JsonSchemaToGeneConverter
+import org.evomaster.core.search.gene.collection.EnumGene
+import org.evomaster.core.search.gene.string.StringGene
 import org.evomaster.core.search.gene.wrapper.CustomMutationRateGene
 import org.evomaster.core.search.gene.wrapper.OptionalGene
-import org.evomaster.core.search.gene.placeholder.CycleObjectGene
-import org.evomaster.core.search.gene.placeholder.LimitObjectGene
-import org.evomaster.core.search.gene.regex.RegexGene
-import org.evomaster.core.search.gene.string.Base64StringGene
-import org.evomaster.core.search.gene.string.StringGene
-import org.evomaster.core.search.gene.utils.GeneUtils
-import org.evomaster.core.search.gene.wrapper.NullableGene
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.net.URI
 import java.net.URISyntaxException
-import java.util.*
-import kotlin.math.max
+import java.util.ArrayDeque
 
 /**
  * https://github.com/OAI/OpenAPI-Specification/blob/3.0.1/versions/3.0.1.md
@@ -74,8 +56,6 @@ object RestActionBuilderV3 {
      * Name given to enum genes representing data examples coming from OpenAPI schema
      */
     const val EXAMPLES_NAME = "SCHEMA_EXAMPLES"
-
-    private val refCache = mutableMapOf<String, Gene>()
 
     /**
      * Key -> schema in the form "name: {...}"
@@ -135,8 +115,42 @@ object RestActionBuilderV3 {
      * clean cache in order to avoid different dto schema with different configurations, eg, enableConstraintHandling.
      */
     fun cleanCache(){
-        refCache.clear()
         dtoCache.clear()
+    }
+
+    /**
+     * Build a [JsonSchemaToGeneConverter] backed by [schemaHolder].
+     * The converter owns its own `$ref` cache, so callers should keep the
+     * returned instance alive for the duration of a single build to benefit
+     * from gene reuse across operations.
+     */
+    private fun newConverter(schemaHolder: RestSchema, options: Options): JsonSchemaToGeneConverter {
+        return JsonSchemaToGeneConverter(
+            RestSchemaRefResolver(schemaHolder, schemaHolder.main),
+            options.toConverterOptions()
+        )
+    }
+
+    private fun newConverter(
+        schemaHolder: RestSchema,
+        currentSchema: SchemaOpenAPI,
+        options: Options
+    ): JsonSchemaToGeneConverter {
+        return JsonSchemaToGeneConverter(
+            RestSchemaRefResolver(schemaHolder, currentSchema),
+            options.toConverterOptions()
+        )
+    }
+
+    private fun Options.toConverterOptions(): JsonSchemaToGeneConverter.Options {
+        return JsonSchemaToGeneConverter.Options(
+            doParseDescription = this.doParseDescription,
+            enableConstraintHandling = this.enableConstraintHandling,
+            invalidData = this.invalidData,
+            probUseDefault = this.probUseDefault,
+            probUseExamples = this.probUseExamples,
+            usingWhiteBox = this.usingWhiteBox
+        )
     }
 
 
@@ -171,6 +185,7 @@ object RestActionBuilderV3 {
         val errorEndpoints = mutableListOf<String>()
 
         val swagger = schemaHolder.main.schemaParsed
+        val converter = newConverter(schemaHolder, options)
 
         swagger.paths
             .forEach { e ->
@@ -182,6 +197,7 @@ object RestActionBuilderV3 {
                     skipped,
                     actionCluster,
                     schemaHolder,
+                    converter,
                     options,
                     errorEndpoints
                 )
@@ -201,6 +217,7 @@ object RestActionBuilderV3 {
         skipped: MutableList<Endpoint>,
         actionCluster: MutableMap<String, Action>,
         schemaHolder: RestSchema,
+        converter: JsonSchemaToGeneConverter,
         options: Options,
         errorEndpoints: MutableList<String>
     ) {
@@ -222,7 +239,7 @@ object RestActionBuilderV3 {
         if (pathItem.`$ref` != null) {
             val pi = SchemaUtils.getReferencePathItem(schemaHolder,schemaHolder.main,pathItem.`$ref`,messages)
                 ?: return
-            handlePathItem(pathKey, pi, messages,endpointsToSkip,skipped,actionCluster,schemaHolder,options,errorEndpoints)
+            handlePathItem(pathKey, pi, messages,endpointsToSkip,skipped,actionCluster,schemaHolder,converter,options,errorEndpoints)
             return
         }
 
@@ -248,6 +265,7 @@ object RestActionBuilderV3 {
                     operation,
                     schemaHolder,
                     schemaHolder.main,
+                    converter,
                     options,
                     errorEndpoints,
                     messages
@@ -310,16 +328,16 @@ object RestActionBuilderV3 {
         val swagger = OpenAPIParser().readContents(schema,null,null)
         val currentSchema = SchemaOpenAPI(schema,swagger.openAPI, SchemaLocation.MEMORY)
         val schemaHolder = RestSchema(currentSchema)
+        val converter = newConverter(schemaHolder, currentSchema, options)
 
         schemas.forEach { (t, u) ->
-            val gene = getGene(t,
-                swagger.openAPI.components.schemas[t]!!,
-                schemaHolder,
-                currentSchema,
-                ArrayDeque(),
+            val gene = converter.getGene(
                 t,
-                options,
-                messages = mutableListOf())
+                swagger.openAPI.components.schemas[t]!!,
+                ArrayDeque(),
+                referenceClassDef = t,
+                messages = mutableListOf()
+            )
             dtoCache[u] = gene
         }
 
@@ -364,16 +382,14 @@ object RestActionBuilderV3 {
         val swagger = OpenAPIParser().readContents(schema,null,null)
         val currentSchema = SchemaOpenAPI(schema,swagger.openAPI, SchemaLocation.MEMORY)
         val schemaHolder = RestSchema(currentSchema)
-        val gene = createObjectGene(
+        val converter = newConverter(schemaHolder, currentSchema, options)
+        val gene = converter.getGene(
             name,
             currentSchema.schemaParsed.components.schemas[name]!!,
-            schemaHolder,
-            currentSchema,
             ArrayDeque(),
-            referenceTypeName,
-            options,
-            listOf(),
-            mutableListOf())
+            referenceClassDef = referenceTypeName,
+            messages = mutableListOf()
+        )
         dtoCache[dtoSchema] = gene
         return gene.copy()
     }
@@ -412,11 +428,16 @@ object RestActionBuilderV3 {
         val swagger = OpenAPIParser().readContents(schema,null,null)
         val currentSchema = SchemaOpenAPI(schema,swagger.openAPI, SchemaLocation.MEMORY)
         val schemaHolder = RestSchema(currentSchema)
+        val converter = newConverter(schemaHolder, currentSchema, options)
 
         unidentified.forEach {s->
-            val gene = getGene(names[s.first],
-                swagger.openAPI.components.schemas[names[s.first]]!!,schemaHolder, currentSchema,
-                ArrayDeque(), referenceTypeNames[s.first], options, messages = mutableListOf())
+            val gene = converter.getGene(
+                names[s.first],
+                swagger.openAPI.components.schemas[names[s.first]]!!,
+                ArrayDeque(),
+                referenceClassDef = referenceTypeNames[s.first],
+                messages = mutableListOf()
+            )
             if (!dtoCache.containsKey(s.second))
                 dtoCache[s.second] = gene
         }
@@ -445,13 +466,14 @@ object RestActionBuilderV3 {
         operation: Operation,
         schemaHolder: RestSchema,
         currentSchema: SchemaOpenAPI,
+        converter: JsonSchemaToGeneConverter,
         options: Options,
         errorEndpoints: MutableList<String>,
         messages: MutableList<String>
     ) {
 
         try{
-            val params = extractParams(verb, restPath, operation, schemaHolder,currentSchema, options, messages)
+            val params = extractParams(verb, restPath, operation, schemaHolder,currentSchema, converter, options, messages)
             repairParams(params, restPath, messages)
 
             val produces = operation.responses?.values //different response objects based on HTTP code
@@ -528,6 +550,7 @@ object RestActionBuilderV3 {
         operation: Operation,
         schemaHolder: RestSchema,
         currentSchema: SchemaOpenAPI,
+        converter: JsonSchemaToGeneConverter,
         options: Options,
         messages: MutableList<String>
     ): MutableList<Param> {
@@ -542,14 +565,14 @@ object RestActionBuilderV3 {
                     if(param == null){
                         messages.add("Failed to handle ${p.`$ref`} in $verb:$restPath")
                     } else {
-                        handleParam(param, schemaHolder, currentSchema, params, options, messages)
+                        handleParam(param, schemaHolder, currentSchema, converter, params, options, messages)
                     }
                 } else {
-                    handleParam(p, schemaHolder,currentSchema, params, options, messages)
+                    handleParam(p, schemaHolder,currentSchema, converter, params, options, messages)
                 }
             }
 
-        handleBodyPayload(operation, verb, restPath, schemaHolder, currentSchema, params, options, messages)
+        handleBodyPayload(operation, verb, restPath, schemaHolder, currentSchema, converter, params, options, messages)
 
         return params
     }
@@ -589,6 +612,7 @@ object RestActionBuilderV3 {
     private fun handleParam(p: Parameter,
                             schemaHolder: RestSchema,
                             currentSchema: SchemaOpenAPI,
+                            converter: JsonSchemaToGeneConverter,
                             params: MutableList<Param>,
                             options: Options,
                             messages: MutableList<String>
@@ -607,13 +631,10 @@ object RestActionBuilderV3 {
             listOf()
         }
 
-        var gene = getGene(
+        var gene = converter.getGene(
             name,
             p.schema,
-            schemaHolder,
-            currentSchema,
             referenceClassDef = null,
-            options = options,
             isInPath = p.`in` == "path",
             examples = examples,
             messages = messages
@@ -678,6 +699,7 @@ object RestActionBuilderV3 {
         restPath: RestPath,
         schemaHolder: RestSchema,
         currentSchema: SchemaOpenAPI,
+        converter: JsonSchemaToGeneConverter,
         params: MutableList<Param>,
         options: Options,
         messages: MutableList<String>
@@ -746,7 +768,7 @@ object RestActionBuilderV3 {
         val deref = obj.schema.`$ref`?.let { ref -> SchemaUtils.getReferenceSchema(schemaHolder, currentSchema, ref, messages) } ?: obj.schema
         val name = deref?.xml?.name ?: deref?.`$ref`?.substringAfterLast("/") ?: "body"
 
-        var gene = getGene(name, obj.schema, schemaHolder,currentSchema, referenceClassDef = null, options = options, messages = messages, examples = examples)
+        var gene = converter.getGene(name, obj.schema, referenceClassDef = null, messages = messages, examples = examples)
 
         if (resolvedBody.required != true && gene !is OptionalGene) {
             gene = OptionalGene(name, gene)
@@ -763,1346 +785,6 @@ object RestActionBuilderV3 {
 
         params.add(bodyParam)
     }
-
-    private fun possiblyOptional(gene: Gene, required: Boolean?): Gene {
-        if (required != true) {
-            return OptionalGene(gene.name, gene).also { GeneUtils.preventCycles(it) }
-        }
-        return gene
-    }
-
-    // internal so the AsyncAPI action builder can reuse the JSON-Schema → Gene
-    // pipeline without duplicating ~800 LOC. A future refactor will lift the
-    // pipeline into a protocol-agnostic JsonSchemaToGeneConverter.
-    internal fun possiblyNullable(gene: Gene, nullable: Boolean) : Gene{
-        if(nullable) {
-            return NullableGene(gene.name, gene).also { GeneUtils.preventCycles(it) }
-        }
-        return gene
-    }
-
-    internal fun getGene(
-        name: String,
-        schema: Schema<*>,
-        schemaHolder: RestSchema,
-        currentSchema: SchemaOpenAPI,
-        history: Deque<String> = ArrayDeque(),
-        referenceClassDef: String?,
-        options: Options,
-        isInPath: Boolean = false,
-        examples: List<Pair<Any,String?>> = listOf(),
-        messages: MutableList<String>
-    ): Gene {
-
-        if (!schema.`$ref`.isNullOrBlank()) {
-            return createObjectFromReference(name, schema.`$ref`, schemaHolder,currentSchema, history, options, examples, messages)
-        }
-
-        /*
-            Nullability is different between 3.0.0 and 3.1.0.
-            Also, it makes sense ONLY for body payloads... as it is not possible to express null values
-            in a URL (eg query and path), as those have ambiguous interpretation, eg,
-            something like /api?x or /api?x= would in most cases treated as an empty string "".
-         */
-        val nullable_3_0 = (schema.nullable != null && schema.nullable)
-        val nullable_3_1 = schema.types?.contains("null") ?: false
-        val nullable = nullable_3_0 || nullable_3_1
-
-        val gene = getNonNullGene(
-            schema,
-            name,
-            options,
-            messages,
-            isInPath,
-            examples,
-            schemaHolder,
-            currentSchema,
-            history,
-            referenceClassDef
-        )
-
-        return possiblyNullable(gene, nullable)
-    }
-
-    internal fun getNonNullGene(
-        schema: Schema<*>,
-        name: String,
-        options: Options,
-        messages: MutableList<String>,
-        isInPath: Boolean,
-        examples: List<Pair<Any, String?>>,
-        schemaHolder: RestSchema,
-        currentSchema: SchemaOpenAPI,
-        history: Deque<String>,
-        referenceClassDef: String?
-    ): Gene {
-        /*
-            https://github.com/OAI/OpenAPI-Specification/blob/3.0.1/versions/3.0.1.md#dataTypeFormat
-
-        Common Name	    type	format	Comments
-        integer	        integer	int32	signed 32 bits
-        long	        integer	int64	signed 64 bits
-        float	        number	float
-        double	        number	double
-        string	        string
-        byte	        string	byte	base64 encoded characters
-        binary	        string	binary	any sequence of octets
-        boolean	        boolean
-        date	        string	date	As defined by full-date - RFC3339
-        dateTime	    string	date-time	As defined by date-time - RFC3339
-        password	    string	password	Used to hint UIs the input needs to be obscured.
-         */
-
-
-        val type = schema.type ?: schema.types?.firstOrNull { it != "null" }
-        val format = schema.format
-
-        if (schema.enum?.isNotEmpty() == true) {
-
-            when (type) {
-                "string" -> {
-                    return EnumGene(name, (schema.enum.map {
-                        if (it !is String)
-                            LoggingUtil.uniqueWarn(
-                                log,
-                                "an item of enum is not string (ie, ${it::class.java.simpleName}) for a property whose `type` is string and `name` is $name"
-                            )
-                        it.toString()
-                    } as MutableList<String>).apply {
-                        if (options.invalidData) {
-                            //Besides the defined values, add one to test robustness
-                            add("EVOMASTER")
-                        }
-                    })
-                        .apply { this.description = schema.description }
-                }
-                /*
-                    Looks like a possible bug in the parser, where numeric enums can be read as strings... got this
-                    issue in GitLab schemas, eg for visibility_level
-                 */
-                "integer" -> {
-                    if (format == "int64") {
-                        val data: MutableList<Long> = schema.enum
-                            .map { if (it is String) it.toLong() else it as Long }
-                            .toMutableList()
-
-                        return EnumGene(name, (data).apply { add(42L) })
-                    }
-
-                    val data: MutableList<Int> = schema.enum
-                        .map { if (it is String) it.toInt() else it as Int }
-                        .toMutableList()
-                    return EnumGene(name, data.apply { add(42) })
-                }
-
-                "number" -> {
-                    //if (format == "double" || format == "float") {
-                    //TODO: Is it always casted as Double even for Float??? Need test
-                    val data: MutableList<Double> = schema.enum
-                        .map { if (it is String) it.toDouble() else it as Double }
-                        .toMutableList()
-                    return EnumGene(name, data.apply { add(42.0) })
-                }
-
-                else -> messages.add("Cannot handle enum of type: $type")
-            }
-        }
-
-
-        //first check for "optional" format
-        when (format?.lowercase()) {
-            "char" -> return buildStringGeneForChar(name, isInPath)
-            "int8", "int16", "int32" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                IntegerGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                format,
-                messages
-            )//IntegerGene(name)
-            "int64" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                LongGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                messages = messages
-            ) //LongGene(name)
-            "double" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                DoubleGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                messages = messages
-            )//DoubleGene(name)
-            "float" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                FloatGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                messages = messages
-            )//FloatGene(name)
-            "password" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                StringGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                messages = messages
-            )//StringGene(name) //nothing special to do, it is just a hint
-            "binary" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                StringGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                messages = messages
-            )//StringGene(name) //does it need to be treated specially?
-            "byte" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                Base64StringGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                messages = messages
-            )//Base64StringGene(name)
-            "date", "local-date" -> return DateGene(name, onlyValidDates = !options.invalidData)
-            "date-time", "local-date-time" -> {
-                val f = if (format?.lowercase() == "date-time") {
-                    FormatForDatesAndTimes.RFC3339
-                } else {
-                    FormatForDatesAndTimes.ISO_LOCAL
-                }
-                return DateTimeGene(
-                    name,
-                    format = f,
-                    date = DateGene("date", onlyValidDates = !options.invalidData, format = f),
-                    time = TimeGene("time", onlyValidTimes = !options.invalidData, format = f)
-                )
-            }
-
-            else -> if (format != null) {
-                messages.add("Unhandled format '$format' for '$name'")
-            }
-        }
-
-        /*
-                If a format is not defined, the type should default to
-                the JSON Schema definition
-         */
-        when (type?.lowercase()) {
-            "integer" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                IntegerGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                messages = messages
-            )//IntegerGene(name)
-            "number" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                DoubleGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                messages = messages
-            )//DoubleGene(name)
-            "boolean" -> return BooleanGene(name)
-            "string" -> {
-                return if (schema.pattern == null) {
-                    createNonObjectGeneWithSchemaConstraints(
-                        schema,
-                        name,
-                        StringGene::class.java,
-                        options,
-                        null,
-                        isInPath,
-                        examples,
-                        messages = messages
-                    ) //StringGene(name)
-                } else {
-                    try {
-                        createNonObjectGeneWithSchemaConstraints(
-                            schema,
-                            name,
-                            RegexGene::class.java,
-                            options,
-                            null,
-                            isInPath,
-                            examples,
-                            messages = messages
-                        )
-                    } catch (e: Exception) {
-                        /*
-                            TODO: if the Regex is syntactically invalid, we should warn
-                            the user. But, as we do not support 100% regex, might be an issue
-                            with EvoMaster. Anyway, in such cases, instead of crashing EM, let's just
-                            take it as a String.
-                            When 100% support, then tell user that it is his/her fault
-                         */
-                        LoggingUtil.uniqueWarn(log, "Cannot handle regex: ${schema.pattern}")
-                        createNonObjectGeneWithSchemaConstraints(
-                            schema,
-                            name,
-                            StringGene::class.java,
-                            options,
-                            null,
-                            isInPath,
-                            examples,
-                            messages = messages
-                        )//StringGene(name)
-                    }
-                }
-            }
-
-            "array" -> {
-                if (schema is ArraySchema || schema is JsonSchema) {
-
-                    val arrayType: Schema<*> = if (schema.items == null) {
-                        LoggingUtil.uniqueWarn(
-                            log, "Array type '$name' is missing mandatory field 'items' to define its type." +
-                                    " Defaulting to 'string'"
-                        )
-                        Schema<Any>().also { it.type = "string" }
-                    } else {
-                        schema.items
-                    }
-
-                    // Use the XML name from schema.xml.name (the name of the array element in XML)
-                    // if available, otherwise fallback to name + "_item"
-                    val itemName = schema.xml?.name ?: (name + "_item")
-                    val template = getGene(
-                        itemName,
-                        arrayType,
-                        schemaHolder,
-                        currentSchema,
-                        history,
-                        referenceClassDef = null,
-                        options = options,
-                        messages = messages
-                    )//Could still have an empty []
-                    //if (template is CycleObjectGene) {
-                    //return CycleObjectGene("<array> ${template.name}")
-                    //}
-                    return createNonObjectGeneWithSchemaConstraints(
-                        schema,
-                        name,
-                        ArrayGene::class.java,
-                        options,
-                        template,
-                        isInPath,
-                        examples,
-                        messages = messages
-                    )//ArrayGene(name, template)
-                } else {
-                    LoggingUtil.uniqueWarn(log, "Invalid 'array' definition for '$name'")
-                }
-            }
-
-            "object" -> {
-                val properties = schema.properties ?: emptyMap()
-
-                val attributeNames = properties
-                    .filterValues { it.xml?.attribute == true }
-                    .keys
-
-                if (attributeNames.isNotEmpty()) {
-                    val fields = properties.map { (propName, propSchema) ->
-                        getGene(
-                            propName,
-                            propSchema,
-                            schemaHolder,
-                            currentSchema,
-                            history,
-                            referenceClassDef,
-                            options,
-                            false,
-                            examples,
-                            messages
-                        )
-                    }
-
-                    return ObjectWithAttributesGene(
-                        name = schema.xml?.name ?: name,
-                        fixedFields = fields,
-                        refType = referenceClassDef,
-                        isFixed = true,
-                        template = null,
-                        additionalFields = null,
-                        attributeNames = attributeNames
-                    )
-                } else {
-                    return createObjectGene(
-                        name,
-                        schema,
-                        schemaHolder,
-                        currentSchema,
-                        history,
-                        referenceClassDef,
-                        options,
-                        examples,
-                        messages
-                    )
-                }
-            }
-            //TODO file is a hack. I want to find a more elegant way of dealing with it (BMR)
-            //FIXME is this even a standard type???
-            "file" -> return createNonObjectGeneWithSchemaConstraints(
-                schema,
-                name,
-                StringGene::class.java,
-                options,
-                null,
-                isInPath,
-                examples,
-                messages = messages
-            ) //StringGene(name)
-        }
-
-        if ((name == "body" || referenceClassDef != null) && schema.properties?.isNotEmpty() == true) {
-            /*
-                name == "body": This could happen when parsing a body-payload as formData
-                referenceClassDef != null : this could happen when parsing a reference of a constraint (eg, anyOf) of the additionalProperties
-            */
-            return createObjectGene(
-                name,
-                schema,
-                schemaHolder,
-                currentSchema,
-                history,
-                referenceClassDef,
-                options,
-                examples,
-                messages
-            )
-        }
-
-        if (type == null && format == null) {
-            return createGeneWithUnderSpecificTypeAndSchemaConstraints(
-                schema, name, schemaHolder, currentSchema, history, referenceClassDef,
-                options, null, isInPath, examples, messages
-            )
-            //createNonObjectGeneWithSchemaConstraints(schema, name, StringGene::class.java, enableConstraintHandling) //StringGene(name)
-        }
-
-        throw IllegalArgumentException("Cannot handle combination $type/$format")
-    }
-
-    /**
-     * @param referenceTypeName is the name of object type
-     */
-    private fun createObjectGene(name: String,
-                                 schema: Schema<*>,
-                                 schemaHolder: RestSchema,
-                                 currentSchema: SchemaOpenAPI,
-                                 history: Deque<String>,
-                                 referenceTypeName: String?,
-                                 options: Options,
-                                 examples: List<Pair<Any,String?>>,
-                                 messages: MutableList<String>
-    ): Gene {
-
-        val fields = schema.properties?.entries?.map {
-            possiblyOptional(
-                getGene(it.key, it.value, schemaHolder,currentSchema, history, referenceClassDef = null, options = options, messages = messages),
-                schema.required?.contains(it.key)
-            )
-        } ?: listOf()
-
-
-        /*
-            additional properties could be
-
-            from OpenAPI Specification v3.1.0 https://spec.openapis.org/oas/latest.html
-            1 - A free-form query parameter, allowing undefined parameters of a specific type:
-                eg, {
-                      "in": "query",
-                      "name": "freeForm",
-                      "schema": {
-                        "type": "object",
-                        "additionalProperties": {
-                          "type": "integer"
-                        },
-                      },
-                      "style": "form"
-                    }
-
-            2 - Model with Map/Dictionary Properties
-                eg, {
-                      "type": "object",
-                      "additionalProperties": {
-                        "type": "string"
-                      }
-                    }
-
-                    {
-                      "type": "object",
-                      "additionalProperties": {
-                        "$ref": "#/components/schemas/ComplexModel"
-                      }
-                    }
-
-             from https://json-schema.org/understanding-json-schema/reference/object.html
-             3 - The additionalProperties keyword is used to control the handling of extra stuff,
-             that is, properties whose names are not listed in the properties keyword
-             or match any of the regular expressions in the patternProperties keyword.
-             By default any additional properties are allowed.
-
-             The value of the additionalProperties keyword is a schema that will be used to validate
-             any properties in the instance that are not matched by properties or patternProperties.
-             Setting the additionalProperties schema to false means no additional properties will be allowed.
-
-             note that the latest version (ie, OpenAPI 3.1.0) does not support patternProperties yet
-             see more with https://docs.readme.com/docs/openapi-compatibility-chart
-         */
-
-        var additionalFieldTemplate : PairGene<StringGene, Gene>? = null
-        var valueTemplate : Gene? = null
-
-        //    Can be either a boolean or a Schema
-        val additional = schema.additionalProperties
-
-        if (additional == null || (additional is Boolean && additional)) {
-            /*
-                if 'false', no other fields besides the specified ones can be added.
-                if 'true', then any kind of extra field can be added
-                Default is 'true' if not specified.
-              */
-            /*
-                FIXME need to fix few things in ObjectGene before we can have this.
-                TODO once fixed, put this back, and reenable failing test in
-                core-it/src/test/kotlin/org/evomaster/core/problem/rest/body/BodyTest.kt
-             */
-//               val name = "valueTemplate"
-//               valueTemplate = ChoiceGene(name, listOf(
-//                   StringGene("${name}_string"),
-//                   BooleanGene("${name}_boolean"),
-//                   IntegerGene("${name}_integer"),
-//                   DoubleGene("${name}_double"),
-//               ))
-        }
-
-        if (additional is Schema<*>) {
-            /*
-               support additionalProperties with schema
-            */
-            if (!additional.`$ref`.isNullOrBlank()) {
-                valueTemplate = createObjectFromReference(
-                    "valueTemplate",
-                    additional.`$ref`,
-                    schemaHolder,
-                    currentSchema,
-                    history,
-                    options = options,
-                    examples = examples,
-                    messages = messages)
-
-            }else if(!additional.type.isNullOrBlank() || additional.types?.isNotEmpty() == true){
-                valueTemplate = getGene(
-                    "valueTemplate",
-                    additional,
-                    schemaHolder,
-                    currentSchema,
-                    history,
-                    null,
-                    options = options,
-                    messages = messages)
-            }
-        }
-        /*
-            TODO could add extra fields for robustness testing,
-            with and without following the given schema for their type
-         */
-
-        if(valueTemplate != null){
-            additionalFieldTemplate= PairGene(
-                "template",
-                StringGene("keyTemplate"),
-                valueTemplate.copy())
-        }
-
-        val attributeNames = schema.properties
-            ?.filter { (_, propSchema) -> propSchema.xml?.attribute == true }
-            ?.map { it.key }
-            ?: emptyList()
-
-        if (attributeNames.isNotEmpty()) {
-            return ObjectWithAttributesGene(
-                name = name,
-                fixedFields = fields,
-                refType = if (schema is ObjectSchema) referenceTypeName ?: schema.title else null,
-                isFixed = true,
-                template = null,
-                additionalFields = null,
-                attributeNames = attributeNames.toSet()
-            )
-        }
-
-        return assembleObjectGeneWithConstraints(
-            name,
-            schema,
-            fields,
-            additionalFieldTemplate,
-            schemaHolder,
-            currentSchema,
-            history,
-            referenceTypeName,
-            options,
-            examples,
-            messages
-        )
-
-    }
-
-    /**
-     *
-     * handled constraints include
-     *      - allOf
-     *      - anyOf
-     *      - oneOf
-     *      - not (OpenAPI not support this yet)
-     */
-    private fun assembleObjectGeneWithConstraints(
-        name: String,
-        schema: Schema<*>,
-        fields: List<Gene>,
-        additionalFieldTemplate: PairGene<StringGene, Gene>?,
-        schemaHolder: RestSchema,
-        currentSchema: SchemaOpenAPI,
-        history: Deque<String>,
-        referenceTypeName: String?,
-        options: Options,
-        examples: List<Pair<Any,String?>>,
-        messages: MutableList<String>
-    ) : Gene{
-        /*
-            TODO discriminator
-            https://spec.openapis.org/oas/latest.html#discriminator-object
-         */
-
-        if (!options.enableConstraintHandling)
-            return assembleObjectGene(name, options, schema, fields, additionalFieldTemplate, referenceTypeName, examples, messages)
-
-        val allOf = schema.allOf?.map { s->
-            //createObjectGene(name, s, swagger, history, null, enableConstraintHandling)
-            getGene(name, s, schemaHolder,currentSchema, history, null, options, messages = messages, examples = examples)
-        }
-
-        val anyOf = schema.anyOf?.map { s->
-            //createObjectGene(name, s, swagger, history, null, enableConstraintHandling)
-            getGene(name, s, schemaHolder,currentSchema, history, null, options, messages = messages, examples = examples)
-        }
-
-        if (!allOf.isNullOrEmpty() && !anyOf.isNullOrEmpty()){
-            messages.add("Cannot handle allOf and oneOf at same time for a schema with name $name")
-            return assembleObjectGene(name, options, schema, fields, additionalFieldTemplate, referenceTypeName, examples, messages)
-        }
-
-        val oneOf = schema.oneOf?.map { s->
-            //createObjectGene(name, s, swagger, history, null, enableConstraintHandling)
-            getGene(name, s, schemaHolder,currentSchema, history, null, options = options, messages = messages)
-        }
-
-        if (!oneOf.isNullOrEmpty() && (!allOf.isNullOrEmpty() || !anyOf.isNullOrEmpty())){
-            messages.add("cannot handle oneOf and allOf/oneOf at same time for a schema with name $name")
-            return assembleObjectGene(name, options, schema, fields, additionalFieldTemplate, referenceTypeName, examples, messages)
-        }
-
-        if (!allOf.isNullOrEmpty()){
-            val allFields = allOf.mapNotNull {
-                when (it) {
-                    is ObjectGene -> it.fields
-                    else -> null
-                }
-            }.flatten()
-            val fields = allFields.plus(fields).distinctBy { it.name }
-            return assembleObjectGene(name, options, schema, fields, additionalFieldTemplate, referenceTypeName, examples, messages)
-        }
-
-        if (!oneOf.isNullOrEmpty()){
-            val choices = if (fields.isEmpty())
-                oneOf
-            else
-                listOf(assembleObjectGene(name, options, schema, fields, additionalFieldTemplate, referenceTypeName, examples, messages)).plus(oneOf)
-
-            return ChoiceGene(name, choices)
-        }
-
-        if (!anyOf.isNullOrEmpty()){
-            val allFields = anyOf.mapNotNull {
-                when (it) {
-                    is ObjectGene -> it.fields
-                    else -> null
-                }
-            }.flatten()
-            /*
-                currently, we handle anyOf as oneOf plus all combined one
-             */
-
-            val choices = if (anyOf.size > 1) {
-                anyOf.plus(
-                    assembleObjectGene(
-                        name, options, schema, allFields.plus(fields).distinctBy { it.name },
-                        additionalFieldTemplate, referenceTypeName, examples, messages
-                    )
-                )
-            } else {
-                anyOf
-            }
-
-            return ChoiceGene(name, choices)
-//            /*
-//                handle all combinations of anyOf
-//                comment it out for the moment
-//             */
-//            return ChoiceGene(name, (1 until 2.0.pow(one.size * 1.0).toInt()).map { i->
-//                assembleObjectGene(
-//                        name,
-//                        schema,
-//                        Integer.toBinaryString(i).toCharArray().mapIndexed { index, c ->
-//                            if (c == '1'){
-//                                one[index].run {
-//                                    when (this) {
-//                                        is ObjectGene -> this.fields
-//                                        is FlexibleObjectGene<*> -> this.fields
-//                                        else -> null
-//                                    }
-//                                }
-//                            }else null
-//                        }.filterNotNull().flatten(),
-//                        additionalFieldTemplate,
-//                        referenceTypeName
-//                )
-//            })
-        }
-
-        return assembleObjectGene(name, options, schema, fields, additionalFieldTemplate, referenceTypeName, examples, messages)
-
-        //TODO not
-    }
-
-    /**
-     * assemble ObjectGene based on [fields] and [additionalFieldTemplate]
-     */
-    private fun assembleObjectGene(
-        name: String,
-        options: Options,
-        schema: Schema<*>,
-        fields: List<Gene>,
-        additionalFieldTemplate: PairGene<StringGene, Gene>?,
-        referenceTypeName: String?,
-        otherExampleValues: List<Pair<Any,String?>>,
-        messages: MutableList<String>
-    ) : Gene{
-        if (fields.isEmpty()) {
-
-            if(options.usingWhiteBox &&
-                schema.additionalProperties == null
-                || (schema.additionalProperties is Boolean && schema.additionalProperties == true)) {
-                //default is true
-                return TaintedMapGene(name, TaintInputName.getTaintName(StaticCounter.getAndIncrease()))
-            }
-
-            if (additionalFieldTemplate != null){
-                return FixedMapGene(name, additionalFieldTemplate)
-            }
-
-            messages.add("No fields for object definition: $name")
-
-            //If we get here, it is really something wrong with the schema...
-            return FixedMapGene(name, PairGene.createStringPairGene(StringGene(name + "_field"), isFixedFirst = true))
-        }
-
-        val mainGene = if (additionalFieldTemplate!=null){
-            ObjectGene(name, fields, if(schema is ObjectSchema) referenceTypeName?:schema.title else null, false, additionalFieldTemplate, mutableListOf())
-        } else {
-            ObjectGene(name, fields, if(schema is ObjectSchema) referenceTypeName?:schema.title else null)
-        }
-
-        val defaultValue = if(options.probUseDefault > 0) schema.default else null
-        val exampleValue = if(options.probUseExamples > 0) schema.example else null
-        val multiExampleValues = if(options.probUseExamples > 0) schema.examples else null
-
-        val examples = mutableListOf<Pair<ObjectGene,String?>>()
-        if(exampleValue != null){
-            duplicateObjectWithExampleFields(name,mainGene, exampleValue)?.let {
-                examples.add(Pair(it,null))
-            }
-        }
-        if(multiExampleValues != null ){
-            examples.addAll(multiExampleValues
-                .mapNotNull { duplicateObjectWithExampleFields(name,mainGene, it) }
-                .map { Pair(it, null) }
-            )
-        }
-        examples.addAll(otherExampleValues
-            .mapNotNull { duplicateObjectWithExampleFields(name,mainGene, it.first)
-                ?.let { obj ->  Pair(obj,it.second) }
-            }
-        )
-
-        val v = examples.map { it.first } //values
-        val n = examples.map{it.second} // names
-
-        val exampleGene = if(examples.isNotEmpty()){
-            ChoiceGene(EXAMPLES_NAME, v, valueNames = n)
-        } else null
-        val defaultGene = if(defaultValue != null){
-            duplicateObjectWithExampleFields("default", mainGene, defaultValue)
-        } else null
-
-        /*
-            add refClass with title of SchemaObject
-            Man: shall we pop history here?
-         */
-        return createGeneWithExampleAndDefault(exampleGene,defaultGene,mainGene,options,name)
-    }
-
-    private fun duplicateObjectWithExampleFields(name: String, mainGene: ObjectGene, exampleValue: Any): ObjectGene? {
-
-        if(exampleValue !is ObjectNode){
-            LoggingUtil.uniqueWarn(log, "When building object example, required an ObjectNode, but found a ${exampleValue.javaClass}")
-            return null
-        }
-
-        val modified = mainGene.fields.map { f ->
-            if(exampleValue.has(f.name)){
-                val e = exampleValue.get(f.name)
-                if(e.isTextual){
-                    EnumGene<String>(f.name, listOf(asRawString(e.textValue())), 0, false)
-                } else if(e.isObject) {
-                    val nested = f.getWrappedGene(ObjectGene::class.java)
-                    if(nested == null){
-                        LoggingUtil.uniqueWarn(log, "When building object example, cannot handle nested object due to gene type: ${f.javaClass}")
-                        f.copy()
-                    } else {
-                        duplicateObjectWithExampleFields(f.name, nested, e)
-                            ?: f.copy()
-                    }
-                } else {
-                    EnumGene<String>(f.name, listOf(""+e.toString()), 0, true)
-                }
-            } else {
-                /*
-                    TODO: if a parameter is optional, and was not specified in the example,
-                    should it be rather skipped? maybe, maybe not... unsure
-                 */
-                f.copy()
-            }
-        }
-        return ObjectGene(
-            name,
-            modified,
-            mainGene.refType,
-            mainGene.isFixed,
-            mainGene.template?.copy() as PairGene<StringGene,Gene>?,
-            mainGene.additionalFields?.map { it.copy() as PairGene<StringGene,Gene>}?.toMutableList()
-        )
-    }
-
-    /**
-     * handle constraints of schema object
-     *
-     * see more
-     * https://spec.openapis.org/oas/v3.0.3#properties
-     * https://spec.openapis.org/oas/latest.html#properties
-     *
-     * handled constraints include
-     *      - minimum (handled by createNonObjectGeneWithSchemaConstraints)
-     *      - maximum (handled by createNonObjectGeneWithSchemaConstraints)
-     *      - exclusiveMaximum (handled by createNonObjectGeneWithSchemaConstraints)
-     *      - exclusiveMinimum (handled by createNonObjectGeneWithSchemaConstraints)
-     *      - maxLength (handled by createNonObjectGeneWithSchemaConstraints)
-     *      - minLength (handled by createNonObjectGeneWithSchemaConstraints)
-     *      - minItems (handled by createNonObjectGeneWithSchemaConstraints)
-     *      - maxItems (handled by createNonObjectGeneWithSchemaConstraints)
-     *      - uniqueItems (handled by createNonObjectGeneWithSchemaConstraints)
-     *      - multipleOf (TODO)
-     *      - allOf (handled by [assembleObjectGeneWithConstraints])
-     *      - anyOf (handled by [assembleObjectGeneWithConstraints])
-     *      - oneOf (handled by [assembleObjectGeneWithConstraints])
-     *      - not (OpenAPI not support this yet)
-     *
-     *
-     * TODO Man might handle example/default property as default later
-     */
-    private fun createGeneWithUnderSpecificTypeAndSchemaConstraints(
-        schema: Schema<*>,
-        name: String,
-        schemaHolder: RestSchema,
-        currentSchema: SchemaOpenAPI,
-        history: Deque<String>,
-        referenceTypeName: String?,
-        options: Options,
-        collectionTemplate: Gene?,
-        isInPath: Boolean,
-        examples: List<Pair<Any,String?>>,
-        messages: MutableList<String>
-    ) : Gene{
-
-        val mightObject = schema.properties?.isNotEmpty() == true || referenceTypeName != null || containsAllAnyOneOfConstraints(schema)
-        if (mightObject){
-            try {
-                return createObjectGene(name, schema, schemaHolder,currentSchema, history, referenceTypeName, options, examples,  messages)
-            }catch (e: Exception){
-                LoggingUtil.uniqueWarn(log, "fail to create ObjectGene for a schema whose `type` and `format` are under specified with error msg: ${e.message?:"no msg"}")
-            }
-        }
-
-        LoggingUtil.uniqueWarn(log, "No type/format information provided for '$name'. Defaulting to 'string'")
-        return createNonObjectGeneWithSchemaConstraints(schema, name, StringGene::class.java, options = options, collectionTemplate,isInPath, listOf(), messages = messages)
-    }
-
-    private fun containsAllAnyOneOfConstraints(schema: Schema<*>) = schema.oneOf != null || schema.anyOf != null || schema.allOf != null
-
-    /**
-     * handled constraints include
-     *      - minimum
-     *      - maximum
-     *      - exclusiveMaximum
-     *      - exclusiveMinimum
-     *      - maxLength
-     *      - minLength
-     *      - minItems
-     *      - maxItems
-     *      - uniqueItems
-     *
-     */
-    private fun createNonObjectGeneWithSchemaConstraints(
-        schema: Schema<*>,
-        name: String,
-        geneClass: Class<*>,
-        options: Options,
-        collectionTemplate: Gene? = null,
-        //might need to add extra constraints if in path
-        isInPath: Boolean,
-        exampleObjects: List<Pair<Any,String?>>,
-        format: String? = null,
-        messages: MutableList<String>
-    ) : Gene{
-
-
-        val mainGene = createMainGene(options, schema, geneClass, format, name, isInPath, collectionTemplate)
-
-        /*
-            See:
-            https://swagger.io/docs/specification/adding-examples/
-            https://swagger.io/specification/
-            https://swagger.io/specification/#schema-object
-
-            TODO This is not a full handling:
-            - example/examples can be defined at same level of "schema" object, ie, in a Parameter Object.
-              "example" behave the same, whereas "examples" is different (here inside "schema" as an array of values,
-              whereas there in a Parameter Object as an array of object definitions).
-            - note that the use of "example" inside a Schema Object is deprecated, and can lead to quite a few unexpected and counter
-              intuitive behavior. should be avoided.
-            - technically there can be "x-example" as well (but that is mainly for older versions of the OpenAPI that
-                did not support example/examples keywords as widely as now?)
-         */
-
-        val defaultValue = if(options.probUseDefault > 0) schema.default else null
-        val exampleValue = if(options.probUseExamples > 0) schema.example else null
-        val multiExampleValues = if(options.probUseExamples > 0) schema.examples else null
-
-        //value and optional name
-        val examples = mutableListOf<Pair<String,String?>>()
-
-        if(exampleValue != null) {
-            val raw = asRawString(exampleValue)
-            examples.add(Pair(raw,null))
-            val arrayM = if(raw.startsWith("[")) "If you are wrongly passing to it an array of values, " +
-                    "the parser would read it as an array string or simply ignore it. "
-            else ""
-            messages.add("The use of 'example' inside a Schema Object is deprecated in OpenAPI. Rather use 'examples'." +
-                    " ${arrayM}Read value: $raw")
-            //TODO a problem here is that currently number arrays would be ignored, and so this message would not written.
-            //however, would need to check if still the case in future in new versions of the parser
-        }
-        if(multiExampleValues != null && multiExampleValues.isNotEmpty()){
-            //possibly bug in parser, but it was reading strings values double-quoted in this case
-            examples.addAll(multiExampleValues.map { Pair(asRawString(it), null) })
-        }
-        examples.addAll( exampleObjects.map { Pair(asRawString(it.first), it.second) })
-
-
-        val defaultGene = if(defaultValue != null){
-            when{
-                NumberGene::class.java.isAssignableFrom(geneClass)
-                    -> EnumGene("default", listOf(defaultValue.toString()),0,true)
-
-                geneClass == StringGene::class.java
-                        || geneClass == Base64StringGene::class.java
-                        || geneClass == RegexGene::class.java
-                    -> EnumGene<String>("default", listOf(asRawString(defaultValue)),0,false)
-
-                //TODO Arrays
-                else -> {
-                    messages.add("Unable to handle 'default': ${asRawString(defaultValue)}")
-                    null
-                }
-            }
-        } else null
-
-        //values
-        val v = examples.map { it.first }
-        //optional names
-        val n = examples.map { it.second }
-
-        val exampleGene = if(examples.isNotEmpty()){
-            when{
-                NumberGene::class.java.isAssignableFrom(geneClass)
-                    -> EnumGene(EXAMPLES_NAME, v,0,true, n)
-
-                geneClass == StringGene::class.java
-                        || geneClass == Base64StringGene::class.java
-                        || geneClass == RegexGene::class.java
-                    -> EnumGene<String>(EXAMPLES_NAME, v,0,false, n)
-
-                //TODO Arrays
-                else -> {
-                    messages.add("Unable to handle 'examples': ${examples.joinToString(" , ")}")
-                    null
-                }
-            }
-        } else null
-
-        return createGeneWithExampleAndDefault(exampleGene, defaultGene, mainGene, options, name)
-    }
-
-    private fun createMainGene(
-        options: Options,
-        schema: Schema<*>,
-        geneClass: Class<*>,
-        format: String?,
-        name: String,
-        isInPath: Boolean,
-        collectionTemplate: Gene?
-    ): Gene {
-
-        /*
-            exclusiveMinimum and exclusiveMaximum differ between schema version 3.0.0 and 3.1.0
-            https://www.openapis.org/blog/2021/02/16/migrating-from-openapi-3-0-to-3-1-0
-
-            Note: in 3.1.0, it is undefined what to do if both minimum and exclusiveMinimum are defined (same issue
-            for max). So, here we ARBITRARILY choose to use only the exclusive constraint, as supporting both
-            would be too much work with practically no meaningful ROI
-         */
-
-        val (minInclusive,minimum) = if(!options.enableConstraintHandling) {
-            Pair(true, schema.exclusiveMinimumValue ?: schema.minimum)
-        } else  if(schema.exclusiveMinimumValue != null){
-            //3.1.0 and above
-            Pair(false,schema.exclusiveMinimumValue)
-        } else {
-            //3.0.0
-            Pair(!(schema.exclusiveMinimum ?: false), schema.minimum)
-        }
-
-        val (maxInclusive, maximum) = if(!options.enableConstraintHandling) {
-            Pair(true, schema.exclusiveMaximumValue ?: schema.maximum)
-        } else  if(schema.exclusiveMaximumValue != null){
-            //3.1.0 and above
-            Pair(false,schema.exclusiveMaximumValue)
-        } else {
-            //3.0.0
-            Pair(!(schema.exclusiveMaximum ?: false),schema.maximum)
-        }
-
-
-        val mainGene = when (geneClass) {
-            // number gene
-            IntegerGene::class.java -> {
-                val minRange: Int
-                val maxRange: Int
-                if (format == "int8") {
-                    minRange = Byte.MIN_VALUE.toInt()
-                    maxRange = Byte.MAX_VALUE.toInt()
-                } else if (format == "int16") {
-                    minRange = Short.MIN_VALUE.toInt()
-                    maxRange = Short.MAX_VALUE.toInt()
-                } else {
-                    minRange = Integer.MIN_VALUE
-                    maxRange = Integer.MAX_VALUE
-                }
-
-                val minConstraint: Int?
-                val maxConstraint: Int?
-                if (options.enableConstraintHandling) {
-                    minConstraint = minimum?.intValueExact()
-                    maxConstraint = maximum?.intValueExact()
-                } else {
-                    minConstraint = null
-                    maxConstraint = null
-                }
-
-                val minValue = if (minConstraint != null) maxOf(minConstraint, minRange) else minRange
-                val maxValue = if (maxConstraint != null) minOf(maxConstraint, maxRange) else maxRange
-
-                IntegerGene(
-                    name,
-                    min = minValue,
-                    max = maxValue,
-                    maxInclusive = maxInclusive,
-                    minInclusive = minInclusive
-                )
-            }
-
-            LongGene::class.java -> LongGene(
-                name,
-                min = if (options.enableConstraintHandling) minimum?.longValueExact() else null,
-                max = if (options.enableConstraintHandling) maximum?.longValueExact() else null,
-                maxInclusive = maxInclusive,
-                minInclusive = minInclusive
-            )
-
-            FloatGene::class.java -> FloatGene(
-                name,
-                min = if (options.enableConstraintHandling) minimum?.toFloat() else null,
-                max = if (options.enableConstraintHandling) maximum?.toFloat() else null,
-                maxInclusive = maxInclusive,
-                minInclusive = minInclusive
-            )
-
-            DoubleGene::class.java -> DoubleGene(
-                name,
-                min = if (options.enableConstraintHandling) minimum?.toDouble() else null,
-                max = if (options.enableConstraintHandling) maximum?.toDouble() else null,
-                maxInclusive = maxInclusive,
-                minInclusive = minInclusive
-            )
-
-            BigDecimalGene::class.java -> BigDecimalGene(
-                name,
-                min = if (options.enableConstraintHandling) minimum else null,
-                max = if (options.enableConstraintHandling) maximum else null,
-                maxInclusive = maxInclusive,
-                minInclusive = minInclusive
-            )
-
-            BigIntegerGene::class.java -> BigIntegerGene(
-                name,
-                min = if (options.enableConstraintHandling) minimum?.toBigIntegerExact() else null,
-                max = if (options.enableConstraintHandling) maximum?.toBigIntegerExact() else null,
-                maxInclusive = maxInclusive,
-                minInclusive = minInclusive
-            )
-            // string, Base64StringGene and regex gene
-            StringGene::class.java -> buildStringGene(name, options, schema, isInPath)
-            Base64StringGene::class.java -> Base64StringGene(name, buildStringGene(name, options, schema, isInPath))
-            RegexGene::class.java -> {
-                /*
-                    TODO handle constraints for regex gene
-                    eg,  min and max
-                    also, isInPath
-                 */
-                RegexHandler.createGeneForEcma262(schema.pattern).apply { this.name = name }
-            }
-
-            ArrayGene::class.java -> {
-                if (collectionTemplate == null)
-                    throw IllegalArgumentException("cannot create ArrayGene when collectionTemplate is null")
-                ArrayGene(
-                    name,
-                    template = collectionTemplate,
-                    uniqueElements = if (options.enableConstraintHandling) schema.uniqueItems ?: false else false,
-                    minSize = if (options.enableConstraintHandling) schema.minItems else null,
-                    maxSize = if (options.enableConstraintHandling) schema.maxItems else null
-                )
-            }
-
-            else -> throw IllegalStateException("cannot create gene with constraints for gene:${geneClass.name}")
-        }
-
-        if (mainGene.description.isNullOrBlank()) {
-            mainGene.description = schema.description
-        }
-        return mainGene
-    }
-
-    private fun createGeneWithExampleAndDefault(
-        exampleGene: Gene?,
-        defaultGene: Gene?,
-        mainGene: Gene,
-        options: Options,
-        name: String
-    ): Gene {
-        if (exampleGene == null && defaultGene == null) {
-            //no special handling
-            return mainGene
-        }
-
-        if (exampleGene != null && defaultGene != null) {
-            val pd = options.probUseDefault
-            val pe = options.probUseExamples
-            val pm = 1 - pd - pe
-            return ChoiceGene(name, listOf(defaultGene, exampleGene, mainGene), 0, listOf(pd, pe, pm))
-        }
-
-        if (exampleGene != null) {
-            val pe = options.probUseExamples
-            val pm = 1 - pe
-            return ChoiceGene(name, listOf(exampleGene, mainGene), 0, listOf(pe, pm))
-        }
-
-        if (defaultGene != null) {
-            val pd = options.probUseDefault
-            val pm = 1 - pd
-            return ChoiceGene(name, listOf(defaultGene, mainGene), 0, listOf(pd, pm))
-        }
-
-        throw IllegalStateException("BUG: logic error, this code should never be reached")
-    }
-
-    private fun asRawString(x : Any) : String {
-        val s = x.toString()
-        if(s.startsWith("\"") && s.endsWith("\""))
-            return s.substring(1, s.length - 1)
-        return s
-    }
-
-    /**
-     * Build a StringGene that represents a char value.
-     * Char values are modelled as string of fixed size 1.
-     */
-    private fun buildStringGeneForChar(
-        name: String,
-        isInPath: Boolean
-    ): StringGene {
-
-        return StringGene(
-            name,
-            minLength = 1,
-            maxLength = 1,
-            invalidChars = if(isInPath) listOf('/','.') else listOf()
-        )
-    }
-
-    private fun buildStringGene(
-        name: String,
-        options: Options,
-        schema: Schema<*>,
-        isInPath: Boolean
-    ): StringGene {
-
-        /*
-            This is bit tricky, as would prevent empty inputs, which are not necessarily invalid.
-            In OpenAPI, path variables are always require.
-            Sending an empty  /users/ for a /users/{id} might or might not work, and could lead to a 405...
-            maybe we skip it, and treat it for a robustness testing
-         */
-        val defaultMin = if(isInPath) 1 else 0
-        //val defaultMin = 0
-
-        return StringGene(
-            name,
-            maxLength = if (options.enableConstraintHandling) schema.maxLength
-                ?: EMConfig.stringLengthHardLimit else EMConfig.stringLengthHardLimit,
-            minLength = max(defaultMin, if (options.enableConstraintHandling) schema.minLength ?: 0 else 0),
-            invalidChars = if(isInPath) listOf('/','.') else listOf()
-        )
-            .apply { this.description = schema.description }
-    }
-
-    internal fun createObjectFromReference(name: String,
-                                          reference: String,
-                                          schemaHolder: RestSchema,
-                                          currentSchema: SchemaOpenAPI,
-                                          history: Deque<String> = ArrayDeque(),
-                                          options: Options,
-                                          examples: List<Pair<Any,String?>>,
-                                          messages: MutableList<String>
-    ): Gene {
-
-        /*
-            The problem in caching objects is that their tree can depend on where they are mounted.
-            For example, assume A->B  will not work (ie cycle) if mounted under another object that has
-            B as ancestor, eg, D->C->B->X->A.
-            An easy case in which this cannot happen is when the target object is a root, ie used directly
-            in a parameter and not inside other objects. In such cases, we can cache it.
-         */
-        val isRoot = history.isEmpty()
-
-        /*
-            We need to avoid cycles like A.B.A...
-            From root to leaves, how many repeated object should appear on a path?
-            TODO Maybe this should be config to experiment with.
-            Anyway, it is a problem in scout-api
-         */
-        val cycleDepth = 1
-
-        if (history.count { it == reference } >= cycleDepth) {
-            return CycleObjectGene("Cycle for: $reference")
-        }
-
-        if (isRoot && refCache.containsKey(reference)) {
-            val copy = refCache[reference]!!.copy()
-            copy.name = name
-            return copy
-        }
-
-        /*
-            TODO This could be customized in EMConfig
-         */
-        val depthLimit = 5
-        if(history.size == depthLimit){
-            return LimitObjectGene("Object-depth limit reached for: $reference")
-        }
-
-        try {
-            //FIXME should not usi URI. see SchemaUtils
-            URI(reference)
-        } catch (e: URISyntaxException) {
-            LoggingUtil.uniqueWarn(log, "Object reference is not a valid URI: $reference")
-        }
-
-        val schema = SchemaUtils.getReferenceSchema(schemaHolder,currentSchema, reference,messages)
-
-        if (schema == null) {
-            //token after last /
-            val classDef = getClassDef(reference)
-
-            LoggingUtil.uniqueWarn(log, "No $classDef among the object definitions in the OpenApi file")
-
-            return ObjectGene(name, listOf(), classDef)
-        }
-
-        history.push(reference)
-
-        val gene = getGene(name, schema, schemaHolder, currentSchema, history, getClassDef(reference), options,  examples = examples, messages = messages)
-
-        if(isRoot) {
-            GeneUtils.preventCycles(gene)
-            GeneUtils.preventLimit(gene)
-            refCache[reference] = gene
-        }
-
-        history.pop()
-
-        return gene
-    }
-
-    private fun getClassDef(reference: String) = reference.substring(reference.lastIndexOf("/") + 1)
 
 
     private fun removeDuplicatedParams(
